Moved Bingo Settings to the Settings folder
